+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.15.4 (Unreleased)
 
 - Fix Alt key being stuck after regaining focus (e. g. after Alt-Tab).
+- Fix blank line height/spacing in print (resolves [issue #41](https://bitbucket.org/cfyzium/bearlibterminal/issues/41/terminal_print-ignores-font-spacing-for)). 
 
 ### 0.15.3 (2017-05-14)